backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
pillow==10.1.0
TTS>=0.20.0
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 